koshly_backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
supabase==2.4.3
google-generativeai==0.7.2
openai==1.30.1
httpx==0.27.0
python-dotenv==1.0.1
razorpay==1.4.1
cryptography==42.0.7
